+++ src/structs/REVMintConfig.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.0;
 /// @custom:member count The number of tokens that should be minted.
 /// @custom:member beneficiary The address that will receive the minted tokens.
 struct REVMintConfig {
-    uint256 chainId;
-    uint256 count;
+    uint16 chainId;
+    uint104 count;
     address beneficiary;
 }
